+++ src/containers/main/TradeHistory/TradeHistory.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useQuery, gql } from '@apollo/client';
+import { useQuery } from '@apollo/client';
 import Skeleton from '@mui/material/Skeleton';
 import Paper from '@mui/material/Paper';
 import Box from '@mui/material/Box';
@@ -20,6 +20,7 @@ import dayjs from 'dayjs';
 import { Coin } from 'constants/coins';
 import styles from './styles.module.scss';
 import { indexIDA } from 'constants/flowConfig';
+import { GET_DISTRIBUTIONS, GET_STREAMS } from './data/queries';
 
 type ColumnName = 'startDate' | 'endDate' | 'Input' | 'Output' | 'PnL';
 
@@ -203,73 +204,6 @@ function EnhancedTableHead(props: EnhancedTableProps) {
 }
 
 const exchangeAddresses = indexIDA.map((ida) => ida.exchangeAddress.toLowerCase());
-
-/**
- * @see https://github.com/Ricochet-Exchange/ricochet-frontend/pull/694#issuecomment-1141946668
- * type 0 or 2 means the stream has been created or terminated, not including updated:
- * @see https://github.com/superfluid-finance/protocol-monorepo/blob/6a151bcd9e4fc6b2b01a838b807320f003900809/packages/sdk-core/src/types.ts#L4-L7
- */
-const GET_STREAMS = gql`
-	query GetStreams($sender: String!, $receivers: [String!]!) {
-		streams: flowUpdatedEvents(
-			where: { sender: $sender, receiver_in: $receivers, type_in: [0, 2] }
-			orderBy: timestamp
-			orderDirection: desc
-			first: 100
-		) {
-			type
-			stream {
-				createdAtTimestamp
-				streamedUntilUpdatedAt
-				updatedAtTimestamp
-				token {
-					symbol
-					id
-				}
-			}
-			receiver
-			timestamp
-			totalAmountStreamedUntilTimestamp
-			transactionHash
-			id
-		}
-	}
-`;
-
-const GET_DISTRIBUTIONS = gql`
-	query GetUserDistributionSubscriptions($subscriber: String!, $updatedAtTimestamps: [String!]!) {
-		distributions: indexSubscriptions(
-			first: 1000
-			where: { subscriber: $subscriber, updatedAtTimestamp_in: $updatedAtTimestamps }
-			orderBy: createdAtTimestamp
-			orderDirection: desc
-		) {
-			id
-			totalAmountReceivedUntilUpdatedAt
-			updatedAtTimestamp
-			createdAtTimestamp
-			approved
-			indexValueUntilUpdatedAt
-			units
-			index {
-				publisher {
-					id
-				}
-				indexValue
-				indexId
-				totalUnitsPending
-				totalUnits
-				updatedAtTimestamp
-				createdAtTimestamp
-				token {
-					id
-					name
-					symbol
-				}
-			}
-		}
-	}
-`;
 
 type TradeHistoryProps = {
 	address: string;
